Декомпозиция проверки на отдельные валидаторы
