+++ Zilon.Core/Zilon.Core.MassSectorGenerator/ISectorValidator.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+
+using LightInject;
+
+using Zilon.Core.Tactics;
+
+namespace Zilon.Core.MassSectorGenerator
+{
+    public interface ISectorValidator
+    {
+        Task Validate(ISector sector, Scope scopeContainer);
+    }
+}
